Make `process_folder` script for general CLI processing
